Refactored TimeInterface to interfaces package
